set up services and shopify config
