+++ src/index.ts
@@ -1,7 +1,6 @@
 // src/index.js
-import express, { Express, Request, Response } from "express";
-import '@shopify/shopify-api/adapters/node';
-import {shopifyApi, LATEST_API_VERSION} from '@shopify/shopify-api';
+import express, { Express, Request, Response, response } from "express";
+import {getStoreInfo} from './services/shopify';
 import dotenv from "dotenv";
 
 dotenv.config();
@@ -9,19 +8,11 @@ dotenv.config();
 const app: Express = express();
 const port = process.env.PORT || 3000;
 
-const shopify = shopifyApi({
-    // The next 4 values are typically read from environment variables for added security
-    apiKey: process.env.apiKey,
-    apiSecretKey: process.env.apiSecretKey || '',
-    hostName: 'localhost:4321',
-    hostScheme: 'http',
-    apiVersion: LATEST_API_VERSION,
-    isEmbeddedApp: true,
-  });
 
-app.get("/", (req: Request, res: Response) => {
-    const client = new shopify.clients.Rest({session});
-    const response = client.get({path: 'shop'});
+app.get("/", async (req: Request, res: Response) => {
+    const store = await getStoreInfo();
+    console.log('Store:', store);
+    return res.status(200).json(store);
 //   res.send(`Express + TypeScript Server + ${apiKey}`);
 });
